CY-14: Restringir acceso desde microservicio de solicitudes
Mejora en estructura y adaptación de test

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/pragma/r2dbc/entity/UserEntity.java
@@ -8,7 +8,6 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -44,7 +43,7 @@ public class UserEntity {
     private String documentNumber;
 
     @Column("base_salary")
-    private BigDecimal baseSalary;
+    private Double baseSalary;
 
     @Column("id_role")
     private Long idRole;
